+++ scripts/after-pack.cjs
@@ -1,0 +1,2 @@
+'use strict'
+module.exports = require('./pty-native-pack.cjs').afterPack